# public/openapi.yaml
openapi: 3.0.0
info:
  title: Api wypoyczalni samochodów
  version: 1.0.0
  description: Api do zarządzania samochodami, klientami i wynajmem
servers:
  - url: http://localhost:8000/api/v1
    description: Środowisko lokalne
components:
  schemas:
    Car:
      type: object
      required:
        - brand
        - model
        - year
        - status
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the car
          readOnly: true
        brand:
          type: string
          example: Toyota
        model:
          type: string
          example: Camry
        year:
          type: integer
          format: int32
          example: 2022
        status:
          type: string
          enum: [available, rented, maintenance]
          example: available
    CarInput:
      type: object
      required:
        - brand
        - model
        - year
        - status
      properties:
        brand:
          type: string
          example: Toyota
        model:
          type: string
          example: Camry
        year:
          type: integer
          format: int32
          example: 2022
        status:
          type: string
          enum: [available, rented, maintenance]
          example: available
    Client:
      type: object
      required:
        - firstName
        - lastName
        - drivingLicenseNumber
        - contact
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the client
          readOnly: true
        firstName:
          type: string
          example: Jan
        lastName:
          type: string
          example: Kowalski
        drivingLicenseNumber:
          type: string
          example: "ABC123456"
        contact:
          type: string
          example: "jan.kowalski@example.com"
    ClientInput:
      type: object
      required:
        - firstName
        - lastName
        - drivingLicenseNumber
        - contact
      properties:
        firstName:
          type: string
          example: Jan
        lastName:
          type: string
          example: Kowalski
        drivingLicenseNumber:
          type: string
          example: "ABC123456"
        contact:
          type: string
          example: "jan.kowalski@example.com"
    Rental:
      type: object
      required:
        - clientId
        - carId
        - startDate
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the rental
          readOnly: true
        clientId:
          type: string
          format: uuid
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        carId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        startDate:
          type: string
          format: date-time
          example: "2024-05-20T10:00:00Z"
        endDate:
          type: string
          format: date-time
          nullable: true
          example: "2024-05-25T10:00:00Z"
        status:
          type: string
          enum: [active, completed]
          example: active
          readOnly: true
    RentalInput:
      type: object
      required:
        - clientId
        - carId
        - startDate
      properties:
        clientId:
          type: string
          format: uuid
          example: "d290f1ee-6c54-4b01-90e6-d701748f0851"
        carId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        startDate:
          type: string
          format: date-time
          example: "2024-05-20T10:00:00Z"
paths:
  /cars:
    get:
      summary: get lista aut
      tags: [Cars]
      responses:
        "200":
          description: Lista aut
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
              example:
                - id: "123e4567-e89b-12d3-a456-426614174000"
                  brand: "Toyota"
                  model: "Corolla"
                  year: 2021
                  status: "available"
    post:
      summary: post dodaj auto
      tags: [Cars]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CarInput"
            example:
              brand: "Ford"
              model: "Focus"
              year: 2023
              status: "available"
      responses:
        "201":
          description: Samochód utworzony
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
              example:
                id: "223e4567-e89b-12d3-a456-426614174001"
                brand: "Ford"
                model: "Focus"
                year: 2023
                status: "available"
        "400":
          description: zle dane
...
  /clients:
    get:
      summary: get lista klientow
      tags: [Clients]
      responses:
        "200":
          description: Lista klientów
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
              example:
                - id: "c790f1ee-6c54-4b01-90e6-d701748f0851"
                  firstName: "Jan"
                  lastName: "Nowak"
                  drivingLicenseNumber: "XYZ987654"
                  contact: "jan.nowak@example.com"
    post:
      summary: Dodaj nowego klienta
      tags: [Clients]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
            example:
              firstName: "Piotr"
              lastName: "Zieliński"
              drivingLicenseNumber: "DEF456123"
              contact: "piotr.zielinski@example.com"
      responses:
        "201":
          description: Klient utworzony
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
              example:
                id: "d290f1ee-6c54-4b01-90e6-d701748f0852"
                firstName: "Piotr"
                lastName: "Zieliński"
                drivingLicenseNumber: "DEF456123"
                contact: "piotr.zielinski@example.com"
        "400":
          description: Błędne dane wejściowe
...
  /rentals:
    get:
      summary: Pobierz aktualne wypożyczenia
      tags: [Rentals]
      responses:
        "200":
          description: Lista aktualnych wypożyczeń
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
              example:
                - id: "rental_id_1"
                  clientId: "d290f1ee-6c54-4b01-90e6-d701748f0851"
                  carId: "123e4567-e89b-12d3-a456-426614174000"
                  startDate: "2024-05-20T10:00:00Z"
                  endDate: null
                  status: "active"
    post:
      summary: Utwórz nowe wypożyczenie
      tags: [Rentals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalInput"
            example:
              clientId: "d290f1ee-6c54-4b01-90e6-d701748f0851"
              carId: "123e4567-e89b-12d3-a456-426614174000"
              startDate: "2024-05-22T14:30:00Z"
      responses:
        "201":
          description: Wypożyczenie utworzone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
              example:
                id: "rental_id_2"
                clientId: "d290f1ee-6c54-4b01-90e6-d701748f0851"
                carId: "123e4567-e89b-12d3-a456-426614174000"
                startDate: "2024-05-22T14:30:00Z"
                endDate: null
                status: "active"
        "400":
          description: Błędne dane wejściowe (np. klient lub samochód nie istnieje, samochód niedostępny)
        "404":
          description: Klient lub samochód nie znaleziony
...